FFmpegreader: Get rid of some invalid casts
The variables here are all signed, and should not be cast to unsigned.
Doing so may result in exceptions.

While removing the cast results in g++ warnings, but it fixes issues with
reading streams where time domain information is not available at the
outset. The warnings are probably harmless.

It's been suggested that the warnings may be caused by questionable
definition of AV_NOPTS_VALUE in FFmpeg codebase, so there is nothing that
(is not hacky) we can do about it here.