tuttleHost: check values that the plugin has returned to the IsIdentity action
The plugin could declare itself as an identity node. In that case, it
returns the name of the input clip and the time to use.
Throw an error:
* if this clip name is not a valid input clip
* if this input time is not available
The input time is not available means that the identity time asked by
the plugin is not contained in the frame range declared by the
framesNeeded action.

Issue #290