tuttleHost: don't put invalid plugins into the plugins cache
We don't want to blacklist a plugin because the load has failed once... maybe just because one guy has launch tuttle with a bad environnement (bad LD_LIBRARY_PATH).
If a plugin can't be loaded, it will not be put into the cache and we will try again at each run.

Fix #292